Added command to build & publish sail runtimes
This helps, as the `laravelsail/*` published on the docker
hub are out of date and not contain all extensions.

If you try to install cashier mollie for example, it won't work
with the image mentioned in the Laravel docs, since it does not
have the intl extension.

As the image that the user downloads for the initial installation
is now very similar to the one for the Laravel service in the
`docker-compose.yml` I can also imagine that this will speed
up the installation process, when no image is found locally.